fix bug

+++ src/main/java/com/hjy/LearnStudy/JavaStudy/Leet/Leet171.java
@@ -10,7 +10,7 @@ public class Leet171 {
     public int titleToNumber(String s) {
         int res = 0;
         for (int i = s.length() - 1; i >= 0; i--) {
-            int nowNumber =(int) (s.charAt(i) - 'A' + 1);
+            int nowNumber = s.charAt(i) - 'A' + 1;
             res += nowNumber * Math.pow(26, s.length() - 1 - i);
         }
         return res;
